messy commit

+++ src/wrappers/ffmpeg/ffmpeg.py
@@ -38,9 +38,13 @@ def trim_video(context: Context, output_file: str):
     subprocess.call(trim_video_command, shell=False, stderr=console_output, stdout=console_output)
 
 
-def convert_video(context: Context, input_file: str, output_file: str):
+def re_encode_video(context: Context, input_file: str, output_file: str, throw_exception = False):
     """
-    Converts a video from input to input with filters needed for d2x to work.
+    Using the "re_encode_video" commands in the yaml to re-encode the input video in an opencv2 friendly
+    manner. Without this step, certain containers might not be compatible with opencv2, and will cause
+    a plethora of errors.
+
+    throw_exception: Will throw a detailed exception and print statement if conversion failed.
     """
     logger = logging.getLogger(__name__)
     frame_rate = context.frame_rate
@@ -50,7 +54,7 @@ def convert_video(context: Context, input_file: str, output_file: str):
                               "-i", input_file]
 
     extract_frames_options = \
-        get_options_from_section(context.config_yaml["ffmpeg"]["video_to_frames"]['output_options'],
+        get_options_from_section(context.config_yaml["ffmpeg"]['re_encode_video']['output_options'],
                                  ffmpeg_command=True)
 
     for element in extract_frames_options:
@@ -58,12 +62,34 @@ def convert_video(context: Context, input_file: str, output_file: str):
 
     extract_frames_command.append("-r")
     extract_frames_command.append(str(frame_rate))
-
     extract_frames_command.extend([output_file])
 
-    console_output = open(context.console_output_dir + "ffmpeg_convert_video.txt", "w")
+    log_file = context.console_output_dir + "ffmpeg_convert_video.txt"
+    console_output = open(log_file, "w")
     console_output.write(str(extract_frames_command))
     subprocess.call(extract_frames_command, shell=False, stderr=console_output, stdout=console_output)
+
+    if throw_exception:
+        with open(context.console_output_dir + "ffmpeg_convert_video.txt") as f:
+            if 'Conversion failed!' in f.read():
+                print("Failed to convert: " + input_file + " -> " + output_file + ".")
+                print("Check the output file for more information: " + log_file)
+
+                raise TypeError
+
+def check_if_file_is_video(ffprobe_dir: str, input_video:str):
+    execute = [
+        ffprobe_dir,
+        "-i", input_video,
+        "-v", "quiet"
+    ]
+
+    return_bytes = subprocess.run(execute, check=True, stdout=subprocess.PIPE).stdout
+
+    if "Invalid data found when processing input" in return_bytes.decode("utf-8"):
+        return False
+
+    return True
 
 
 def extract_frames(context: Context, input_file: str):
@@ -174,7 +200,7 @@ def concat_encoded_vids(context: Context, output_file: str):
     subprocess.call(concat_videos_command, shell=False, stderr=console_output, stdout=console_output)
 
 
-def migrate_tracks(context: Context, no_audio: str, file_dir: str, output_file: str):
+def migrate_tracks(context: Context, no_audio: str, file_dir: str, output_file: str, copy_if_failed = False):
     """
     Add the audio tracks from the original video to the output video.
     """
@@ -202,20 +228,20 @@ def migrate_tracks(context: Context, no_audio: str, file_dir: str, output_file: 
     console_output.write(str(migrate_tracks_command))
     subprocess.call(migrate_tracks_command, shell=False, stderr=console_output, stdout=console_output)
 
-    with open(context.console_output_dir + "migrate_tracks_command.txt") as f:
-        if 'Conversion failed!' in f.read():
-            import os
-            import shutil
+    if copy_if_failed:
+        with open(context.console_output_dir + "migrate_tracks_command.txt") as f:
+            if 'Conversion failed!' in f.read():
+                import os
+                import shutil
 
-            print("conversion failed in migration for some reason, letting this exist as technical debt for now. ")
-            os.remove(output_file)
-            shutil.copy(no_audio, output_file)
+                print("Migrating Tracks failed... copying video in order to continue with dandere2x.")
+                os.remove(output_file)
+                shutil.copy(no_audio, output_file)
 
 def concat_two_videos(context: Context, video_1: str, video_2: str, output_video: str):
     # load context
 
     workspace = context.workspace
-
     temp_concat_file = workspace + "concat_list.txt"
 
     file = open(temp_concat_file, "a")
